fix importing and pathing after minor restructure
