+++ curriculum_navigation/src/TurtleBot_crafting/TurtleBot_v0/envs/utils/EnvironmentHandlers.py
@@ -14,7 +14,7 @@ import rospy
 import sys
 
 
-class TurtleBotRosNode:
+class TurtleBotRosNode(object):
     def __init__(self, timeout_seconds=15):
         rospy.init_node("TurtleBotCurriculumNav", anonymous=False)
